pyproject.toml
[tool.poetry]
name = "cloud-security-scanner"
version = "1.0.0"
description = "Scanner de segurança multicloud — Security Groups, SOs e contêineres desatualizados"
authors = ["Ivonaldo Micheluti Dias <ivonaldo.dias@hotmail.com>"]
readme = "README.md"
license = "MIT"
keywords = ["security", "cloud", "aws", "oci", "gcp", "devsecops", "finops"]

[tool.poetry.dependencies]
python        = "^3.11"
boto3         = "^1.34"
python-dotenv = "^1.0"
rich          = "^13.7"

[tool.poetry.group.dev.dependencies]
pytest        = "^8.0"
pytest-cov    = "^5.0"
mypy          = "^1.9"
black         = "^24.0"
ruff          = "^0.4"
boto3-stubs   = {extras = ["ec2", "ecs", "ecr"], version = "^1.34"}

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts   = "-v --cov=scanners --cov=scoring --cov=reporter --cov-report=term-missing"

[tool.mypy]
python_version      = "3.11"
strict              = true
warn_return_any     = true
warn_unused_configs = true

[tool.black]
line-length    = 100
target-version = ["py311"]

[tool.ruff]
line-length    = 100
target-version = "py311"
select         = ["E", "F", "I", "N", "W", "UP"]

[build-system]
requires      = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
